% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etape1.R
\name{Etape1_traitement_initial}
\alias{Etape1_traitement_initial}
\title{Traitement initial de la base laitier}
\usage{
Etape1_traitement_initial(path)
}
\arguments{
\item{path}{Chemin vers le fichier .dta (base produit laitier)}
}
\value{
Un dataframe nettoyé et enrichi avec les colonnes d'incohérences
}
\description{
Cette fonction lit une base .dta, renomme, labellise, vérifie les identifiants et détecte les incohérences
}
