% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etape2.R
\name{Etape2_analyse_valeurs_unitaires_et_sources}
\alias{Etape2_analyse_valeurs_unitaires_et_sources}
\title{Analyse des valeurs unitaires et classification par source}
\usage{
Etape2_analyse_valeurs_unitaires_et_sources(data)
}
\arguments{
\item{data}{Dataframe contenant au moins : produit, unite_dernier_achat, taille_dernier_achat, VU,
quantite_consommee, qte_autoconsommee, qte_recue}
}
\value{
Un dataframe enrichi avec les colonnes QuantitéConsommée, ValeurConsommée, source
}
\description{
À partir d'un jeu de données contenant la variable VU (valeur unitaire), calcule les statistiques
descriptives par produit/unité/taille, choisit une valeur unitaire représentative (mode ou médiane),
calcule la valeur consommée par ménage, et classe la source de consommation.
}
