% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etape3.R
\name{Etape3_conversion_standardisation}
\alias{Etape3_conversion_standardisation}
\title{Conversion des quantités en unités standard (kg) avec gestion des poids manquants}
\usage{
Etape3_conversion_standardisation(data, baseVUmenage, chemin_conversion)
}
\arguments{
\item{data}{Dataframe contenant les données laitières.}

\item{baseVUmenage}{Dataframe avec valeurs unitaires consolidées.}

\item{chemin_conversion}{Chemin vers le fichier Excel de conversion.}
}
\value{
Dataframe fusionnée avec quantités standardisées.
}
\description{
Cette fonction fait la jointure avec la table de conversion, ajuste les poids manquants,
et calcule les quantités standardisées en kilogrammes.
}
