% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etape4.R
\name{Etape4_nettoyage_validation}
\alias{Etape4_nettoyage_validation}
\title{Nettoyage, détection d’outliers, validation et enrichissement des données}
\usage{
Etape4_nettoyage_validation(
  Base_X1_SemiApurée,
  data_principal,
  mon_path,
  ehcvm_all
)
}
\arguments{
\item{Base_X1_SemiApurée}{Dataframe semi-apuré avec quantités et valeurs.}

\item{data_principal, }{Chemin du dossier contenant les fichiers de données annexes.}

\item{mon_path, }{Chemin du dossier contenant les fichiers de données annexes.}

\item{ehcvm_all, }{Chemin du dossier contenant les fichiers de données annexes}
}
\value{
Dataframe nettoyée et enrichie prête pour analyse.
}
\description{
Cette fonction réalise la détection des valeurs aberrantes, la validation par calories,
et l’ajout des informations géographiques et démographiques.
}
