% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etape5.R
\name{Etape5_calcul_indicateurs}
\alias{Etape5_calcul_indicateurs}
\title{Calcul des indicateurs de consommation laitière}
\usage{
Etape5_calcul_indicateurs(Base_X1_Apurée, data)
}
\arguments{
\item{Base_X1_Apurée}{Dataframe contenant les données semi-apurées.}

\item{data}{Dataframe source contenant les données brutes.}
}
\value{
Liste de dataframes avec les indicateurs.
}
\description{
Cette fonction calcule plusieurs indicateurs descriptifs à partir des bases fournies.
}
